+++ packages/ui/src/lib/index.ts
@@ -54,14 +54,14 @@ export { default as SidebarGroupTitle } from './sidebar/elements/sidebarSection/
 export { default as SidebarSection } from './sidebar/elements/sidebarSection/SidebarSection.svelte';
 export { default as SidebarSectionTitle } from './sidebar/elements/sidebarSection/sidebarSectionTitle/SidebarSectionTitle.svelte';
 
-export { default as SidebarTabLabel } from './sidebar/elements/sidebarTabs/SidebarTabLabel.svelte';
 export { default as SidebarTabList } from './sidebar/elements/sidebarTabs/SidebarTabList.svelte';
 
-export { default as SidebarFromArray } from './sidebar/SidebarFromArray.svelte';
+export { default as LoadingIndicator } from './loadingIndicator/LoadingIndicator.svelte';
 
 export { default as Spinner } from './spinners/Spinner.svelte';
 export { default as TabLabel } from './tabs/TabLabel.svelte';
 export { default as TabList } from './tabs/TabList.svelte';
+export { default as TabPanel } from './tabs/TabPanel.svelte';
 
 export { default as Tooltip } from './tooltip/Tooltip.svelte';
 
@@ -84,7 +84,7 @@ export { default as PlaceholderImage } from './placeholderImage/PlaceholderImage
 
 export { default as Theme } from './theme/Theme.svelte';
 export * from './theme/themeStore';
-export { default as themeSwitcher } from './theme/ThemeSwitcher.svelte';
+export { default as ThemeSwitcher } from './theme/ThemeSwitcher.svelte';
 
 export * from './uniformInput/types';
 export { default as UniformInput } from './uniformInput/UniformInput.svelte';
@@ -97,7 +97,15 @@ export { default as SchemaForm } from './forms/SchemaForm.svelte';
 export * from './forms/types';
 
 export { default as MergeValuesControl } from './mergeValuesControl/MergeValuesControl.svelte';
+
 export { default as Overlay } from './overlay/Overlay.svelte';
+export { default as Trigger } from './overlay/Trigger.svelte';
+
+export * from './auth/auth';
+export { default as AuthMenu } from './auth/AuthMenu.svelte';
+export { default as HandleRedirectFromAuth } from './auth/HandleRedirectFromAuth.svelte';
+export { default as LoginForm } from './auth/LoginForm.svelte';
+export { default as LoginLink } from './auth/LoginLink.svelte';
 
 export { classNames } from './utils/classNames';
 export { randomId } from './utils/randomId';