+++ packages/ui/src/lib/layerControl/LayerContolGroup.stories.svelte
@@ -10,15 +10,14 @@
 
 <script lang="ts">
 	let optionsForGroup = [
-		{ id: 'bus', label: 'Bus stops', colorName: 'data.categorical.blue' },
+		{ id: 'bus', label: 'Bus stops' },
 		{
 			id: 'train',
 			label: 'Train stations',
-			colorName: 'data.categorical.green',
 			hint: 'Excluding underground stations'
 		},
-		{ id: 'underground', label: 'Underground stations', colorName: 'data.categorical.darkpink' },
-		{ id: 'taxi', label: 'Taxi ranks', colorName: 'data.categorical.orange', disabled: true }
+		{ id: 'underground', label: 'Underground stations' },
+		{ id: 'taxi', label: 'Taxi ranks', disabled: true }
 	];
 
 	let optionsForGroup2 = [
@@ -31,8 +30,8 @@
 		{
 			id: 'underground',
 			label: 'Underground stations',
-			hideOpacityControl: true,
-			hideSizeControl: true
+			disableOpacityControl: true,
+			disableSizeControl: true
 		}
 	];
 
@@ -58,20 +57,26 @@
 		},
 		taxi: {
 			colorName: 'data.categorical.orange',
-			visible: true,
+			visible: false,
 			opacity: 1.0,
 			size: 1
 		}
 	};
 	let state2 = {
 		bus: {
-			color: '#00AEEF',
+			colorName: 'data.categorical.blue',
+			visible: true,
+			opacity: 1.0,
+			size: 1
+		},
+		train: {
+			colorName: 'data.categorical.green',
 			visible: true,
 			opacity: 1.0,
 			size: 1
 		},
 		underground: {
-			color: '#9E0059',
+			colorName: 'data.categorical.darkpink',
 			visible: true,
 			opacity: 1.0,
 			size: 1
@@ -86,21 +91,59 @@
 
 <Story name="Default" source />
 
-<Story name="Hide controls size and opacity controls">
-	<div class="space-y-4">
-		<LayerControlGroup
-			bind:options={optionsForGroup}
-			bind:state={state1}
-			hideOpacityControl
-			hideSizeControl
-		/>
-	</div>
+<Story name="With label">
+	<LayerControlGroup
+		bind:options={optionsForGroup}
+		bind:state={state1}
+		disableOpacityControl
+		disableSizeControl
+		label="Layer Control Group"
+	/>
+
 	<pre class="mt-4 text-xs">{JSON.stringify(state1, null, 2)}</pre>
 </Story>
 
-<Story name="Hide controls size and opacity controls for some layers">
-	<LayerControlGroup bind:options={optionsForGroup2} bind:state={state2} />
+<Story name="With hint">
+	<LayerControlGroup
+		bind:options={optionsForGroup}
+		bind:state={state1}
+		disableOpacityControl
+		disableSizeControl
+		label="Layer Control Group"
+		hint="Turn the layers of the map on and off"
+	/>
+
 	<pre class="mt-4 text-xs">{JSON.stringify(state1, null, 2)}</pre>
+</Story>
+
+<Story name="With description">
+	<LayerControlGroup
+		bind:options={optionsForGroup}
+		bind:state={state1}
+		disableOpacityControl
+		disableSizeControl
+		label="Layer Control Group"
+		hint="Turn the layers of the map on and off"
+		description="Transport layers - Taxis disabled"
+	/>
+
+	<pre class="mt-4 text-xs">{JSON.stringify(state1, null, 2)}</pre>
+</Story>
+
+<Story name="Hide controls size and opacity controls">
+	<LayerControlGroup
+		bind:options={optionsForGroup}
+		bind:state={state1}
+		disableOpacityControl
+		disableSizeControl
+	/>
+
+	<pre class="mt-4 text-xs">{JSON.stringify(state1, null, 2)}</pre>
+</Story>
+
+<Story name="Disable controls size and opacity controls for some layers">
+	<LayerControlGroup bind:options={optionsForGroup2} bind:state={state2} />
+	<pre class="mt-4 text-xs">{JSON.stringify(state2, null, 2)}</pre>
 </Story>
 
 <Story name="In small text size context">
@@ -108,8 +151,8 @@
 		<LayerControlGroup
 			bind:options={optionsForGroup}
 			bind:state={state1}
-			hideOpacityControl
-			hideSizeControl
+			disableOpacityControl
+			disableSizeControl
 		/>
 	</div>
 </Story>
@@ -124,6 +167,27 @@ For example, choropleth layers would cover each other.
 		bind:state={state1}
 		mutuallyExclusive
 		name="mutually-exclusive-layers"
+	/>
+
+	<pre class="mt-4 text-xs">{JSON.stringify(state1, null, 2)}</pre>
+</Story>
+
+<Story name="Disabled (global)">
+	<LayerControlGroup bind:options={optionsForGroup} bind:state={state1} disabled name="Disabled" />
+
+	<pre class="mt-4 text-xs">{JSON.stringify(state1, null, 2)}</pre>
+</Story>
+
+<Story name="With error">
+	<LayerControlGroup
+		bind:options={optionsForGroup}
+		bind:state={state1}
+		label="Preferred mode of transport"
+		hint="Contextual Hint"
+		description="Deselect all to see an error state!"
+		error={Object.values(state1).every((layer) => layer.visible === false)
+			? 'You must select an option'
+			: undefined}
 	/>
 
 	<pre class="mt-4 text-xs">{JSON.stringify(state1, null, 2)}</pre>