+++ packages/ui/src/lib/layerControl/LayerControl.svelte
@@ -11,20 +11,22 @@
 	import ColorPicker from './ColorPicker.svelte';
 	import ResizeControl from './ResizeControl.svelte';
 
+	import { randomId } from '../utils/randomId';
+
 	/**
 	 * if `true`, then the trigger to open the color picker is not displayed
 	 */
-	export let hideColorControl = false;
+	export let disableColorControl = false;
 
 	/**
 	 * if `true`, then the trigger to open the opacity control is not displayed
 	 */
-	export let hideOpacityControl = false;
+	export let disableOpacityControl = false;
 
 	/**
 	 * if `true`, then the trigger to open the size control is not displayed
 	 */
-	export let hideSizeControl = false;
+	export let disableSizeControl = false;
 
 	/**
 	 * the name of the layer
@@ -32,9 +34,16 @@
 	export let label = '';
 
 	/**
-	 * (optional) explanatory help text to be displayed in tooltip
+	 * Optional help text that appears in a tooltip when a user interacts with the tooltip trigger.
+	 * It provides additional information intended to help the user decide whether to check the checkbox.
 	 */
 	export let hint = '';
+
+	/**
+	 * Optional text that appears next to the information icon (the letter "i" in a circle) in the tooltip trigger.
+	 * It provides additional clues that help text is available (e.g. "More information", "About", "Help")
+	 */
+	export let hintLabel = '';
 
 	export let disabled = false;
 
@@ -77,9 +86,9 @@
 	 */
 	export let selectedOptionId: string | undefined = undefined;
 	/**
-	 * Id of this option  (used only if `mutuallyExclusive` is true).
+	 * Id of this option.
 	 */
-	export let optionId = '';
+	export let optionId = randomId();
 
 	/**
 	 * Name of the radio button group  (used only if `mutuallyExclusive` is true)
@@ -87,9 +96,10 @@
 	export let name = '';
 
 	/**
-	 * text that appears in the hint tooltip target, next to the icon
+	 * List of controls for which placeholder should be displayed in control is hidden.
+	 * This enables alignment of controls between different `LayerControl`s with different controls enabled,
 	 */
-	export let hintLabel = '';
+	export let controlsInUse: ('color' | 'opacity' | 'size')[] = ['color', 'opacity', 'size'];
 </script>
 
 <div class="flex items-center space-x-1">
@@ -104,27 +114,53 @@
 				{name}
 			/>
 		{:else}
-			<Checkbox bind:checked={state.visible} label="" {disabled} />
+			<Checkbox bind:checked={state.visible} label="" {disabled} id={optionId} />
 		{/if}
 	</div>
 
-	{#if !hideColorControl}
-		<ColorPicker bind:colorName={state.colorName} />
+	{#if controlsInUse.includes('color')}
+		<ColorPicker
+			bind:colorName={state.colorName}
+			disabled={disabled || disableColorControl}
+			{label}
+		/>
 	{/if}
 
-	{#if !hideOpacityControl}
-		<OpacityControl bind:opacity={state.opacity} />
+	{#if controlsInUse.includes('opacity')}
+		<OpacityControl
+			bind:opacity={state.opacity}
+			disabled={disabled || disableOpacityControl}
+			{label}
+		/>
 	{/if}
 
-	{#if !hideSizeControl}
-		<ResizeControl bind:size={state.size} {minSize} {maxSize} />
+	{#if controlsInUse.includes('size')}
+		<ResizeControl
+			bind:size={state.size}
+			{minSize}
+			{maxSize}
+			disabled={disabled || disableSizeControl}
+			{label}
+		/>
 	{/if}
+
 	{#if label}
-		<span class="form-label font-normal leading-none">{label}</span>
+		<label class="form-label font-normal leading-none" for={optionId}>{label}</label>
 	{/if}
+
+	{#if $$slots.hint}
+		<!-- An optional `<Overlay>` component to provide additional explanation. -->
+		<slot name="hint" />
+	{/if}
+
 	{#if hint}
 		<Overlay>
-			<Trigger slot="trigger" size="xs" {hintLabel} />
+			<Trigger
+				slot="trigger"
+				size="xs"
+				{hintLabel}
+				ariaLabel={!hintLabel && label ? label : null}
+			/>
 			{hint}
 		</Overlay>
 	{/if}