+++ packages/ui/src/lib/sidebar/elements/sidebarSection/sidebarGroupTitle/SidebarGroupTitle.svelte
@@ -5,14 +5,23 @@
 	 */
 </script>
 
-<div class="flex justify-between items-end text-color-text-primary">
-	<h3 class="font-semibold text-sm leading-snug">
-		<!-- The title. -->
-		<slot />
-	</h3>
+<div class=" text-color-text-primary">
+	<div class="flex justify-between items-end">
+		<h3 class="font-semibold text-sm leading-snug">
+			<!-- The title. -->
+			<slot />
+		</h3>
 
-	{#if $$slots.hint}
-		<!-- An optional `<Overlay>` component to provide additional explanation. -->
-		<slot name="hint" />
+		{#if $$slots.hint}
+			<!-- An optional `<Overlay>` component to provide additional explanation. -->
+			<slot name="hint" />
+		{/if}
+	</div>
+
+	{#if $$slots.subTitle}
+		<div class="text-xs my-1">
+			<!-- Optional longer subtitle to display below the main title. -->
+			<slot name="subTitle" />
+		</div>
 	{/if}
 </div>